src/requirements.txt
asgiref==3.8.1
Django==4.2.17
gpg==1.15.1
kafka-python==2.0.2
libcomps==0.1.20
rpm==4.16.1.3
sqlparse==0.5.3
typing_extensions==4.12.2
